// src/interface.ts
export interface Actividad {
    activity:      string;
    availability:  number;
    type:          string;
    participants:  number;
    price:         number;
    accessibility: string;
    duration:      string;
    kidFriendly:   boolean;
    link:          string;
    key:           string;
}